xxxxxxxxxxxxxxxxx

+++ trunk/hlfs/src/clean/Mapreducer/mr_clean_route.cpp
@@ -56,9 +56,9 @@ class SegUsageCalcMap: public HadoopPipes::Mapper {
             HADOOP_ASSERT(m_storage != NULL, "failed to init storage handler ");
             uint32_t segment_size;
             uint32_t block_size;
-            uint32_t max_fs_size;
+            uint64_t max_fs_size;
             int ret = read_fs_meta(m_storage, &segment_size, &block_size,&max_fs_size);
-            printf("DBG:--segment size:%u,block size:%u,max fs size%u\n",segment_size,block_size,max_fs_size);
+            printf("DBG:--segment size:%u,block size:%u,max fs size%llu\n",segment_size,block_size,max_fs_size);
             //m_latest_inode = load_latest_inode(m_storage); 
      	    //HADOOP_ASSERT(m_latest_inode != NULL, "failed to load latest inode ");
             m_block_size = block_size;
