+++ trunk/hlfs/src/tools/hlfs_seg_usage_calc.c
@@ -58,7 +58,7 @@ int main(int argc, char *argv[])
     g_message("storage init over....");
     uint32_t segment_size;
     uint32_t block_size;
-    uint32_t max_fs_size;
+    uint64_t max_fs_size;
     int ret = read_fs_meta(storage, &segment_size, &block_size,&max_fs_size);
     g_message("segment size:%d,block size:%d,max fs size%d",segment_size,block_size,max_fs_size);
 #if 0
